Move rst_validator.py to tests/, change import in test_rstwriter.py to match
